+++ src/main/java/Wholesale/SpecialPrice.java
@@ -5,10 +5,10 @@ import Stock.Product;
 public class SpecialPrice {
 
     private Product product;    //Has-A relationship through composition
-    private float discountPrice;
+    private int discountPrice;
     private int noOfItems;
 
-    public SpecialPrice(Product product, int noOfItems, float discountPrice) {
+    public SpecialPrice(Product product, int noOfItems, int discountPrice) {
         this.product = product;
         this.discountPrice = discountPrice;
         this.noOfItems = noOfItems;
@@ -16,7 +16,7 @@ public class SpecialPrice {
 
     public Product getStockItem() { return product; }
 
-    public float getDiscountPrice() { return discountPrice; }
+    public int getDiscountPrice() { return discountPrice; }
 
     public int getNoOfItems() { return noOfItems; }
 }
